started to clean pytests
